split lidarr into more files
