+++ lidarr/command.go
@@ -1,0 +1,37 @@
+package lidarr
+
+import (
+	"encoding/json"
+	"fmt"
+)
+
+// GetCommands returns all available Lidarr commands.
+func (l *Lidarr) GetCommands() ([]*CommandResponse, error) {
+	var output []*CommandResponse
+
+	if err := l.GetInto("v1/command", nil, &output); err != nil {
+		return nil, fmt.Errorf("api.Get(command): %w", err)
+	}
+
+	return output, nil
+}
+
+// SendCommand sends a command to Lidarr.
+func (l *Lidarr) SendCommand(cmd *CommandRequest) (*CommandResponse, error) {
+	var output CommandResponse
+
+	if cmd == nil || cmd.Name == "" {
+		return &output, nil
+	}
+
+	body, err := json.Marshal(cmd)
+	if err != nil {
+		return nil, fmt.Errorf("json.Marshal(cmd): %w", err)
+	}
+
+	if err := l.PostInto("v1/command", nil, body, &output); err != nil {
+		return nil, fmt.Errorf("api.Post(command): %w", err)
+	}
+
+	return &output, nil
+}
