+++ lidarr/qualityprofile.go
@@ -1,0 +1,53 @@
+package lidarr
+
+import (
+	"encoding/json"
+	"fmt"
+	"strconv"
+
+	"golift.io/starr"
+)
+
+// GetQualityProfiles returns the quality profiles.
+func (l *Lidarr) GetQualityProfiles() ([]*QualityProfile, error) {
+	var profiles []*QualityProfile
+
+	err := l.GetInto("v1/qualityprofile", nil, &profiles)
+	if err != nil {
+		return nil, fmt.Errorf("api.Get(qualityprofile): %w", err)
+	}
+
+	return profiles, nil
+}
+
+// AddQualityProfile updates a quality profile in place.
+func (l *Lidarr) AddQualityProfile(profile *QualityProfile) (int64, error) {
+	post, err := json.Marshal(profile)
+	if err != nil {
+		return 0, fmt.Errorf("json.Marshal(profile): %w", err)
+	}
+
+	var output QualityProfile
+
+	err = l.PostInto("v1/qualityProfile", nil, post, &output)
+	if err != nil {
+		return 0, fmt.Errorf("api.Post(qualityProfile): %w", err)
+	}
+
+	return output.ID, nil
+}
+
+// UpdateQualityProfile updates a quality profile in place.
+func (l *Lidarr) UpdateQualityProfile(profile *QualityProfile) error {
+	put, err := json.Marshal(profile)
+	if err != nil {
+		return fmt.Errorf("json.Marshal(profile): %w", err)
+	}
+
+	_, err = l.Put("v1/qualityProfile/"+strconv.FormatInt(profile.ID, starr.Base10), nil, put)
+	if err != nil {
+		return fmt.Errorf("api.Put(qualityProfile): %w", err)
+	}
+
+	return nil
+}
