+++ lidarr/system.go
@@ -1,0 +1,31 @@
+package lidarr
+
+import (
+	"fmt"
+
+	"golift.io/starr"
+)
+
+// GetSystemStatus returns system status.
+func (l *Lidarr) GetSystemStatus() (*SystemStatus, error) {
+	var status SystemStatus
+
+	err := l.GetInto("v1/system/status", nil, &status)
+	if err != nil {
+		return nil, fmt.Errorf("api.Get(system/status): %w", err)
+	}
+
+	return &status, nil
+}
+
+// GetBackupFiles returns all available Lidarr backup files.
+// Use GetBody to download a file using BackupFile.Path.
+func (l *Lidarr) GetBackupFiles() ([]*starr.BackupFile, error) {
+	var output []*starr.BackupFile
+
+	if err := l.GetInto("v1/system/backup", nil, &output); err != nil {
+		return nil, fmt.Errorf("api.Get(system/backup): %w", err)
+	}
+
+	return output, nil
+}
