+++ lidarr/tag.go
@@ -1,0 +1,50 @@
+package lidarr
+
+import (
+	"encoding/json"
+	"fmt"
+
+	"golift.io/starr"
+)
+
+// GetTags returns all the tags.
+func (l *Lidarr) GetTags() ([]*starr.Tag, error) {
+	var tags []*starr.Tag
+
+	err := l.GetInto("v1/tag", nil, &tags)
+	if err != nil {
+		return nil, fmt.Errorf("api.Get(tag): %w", err)
+	}
+
+	return tags, nil
+}
+
+// AddTag adds a tag or returns the ID for an existing tag.
+func (l *Lidarr) AddTag(label string) (int, error) {
+	body, err := json.Marshal(&starr.Tag{Label: label, ID: 0})
+	if err != nil {
+		return 0, fmt.Errorf("json.Marshal(tag): %w", err)
+	}
+
+	var tag starr.Tag
+	if err = l.PostInto("v1/tag", nil, body, &tag); err != nil {
+		return tag.ID, fmt.Errorf("api.Post(tag): %w", err)
+	}
+
+	return tag.ID, nil
+}
+
+// UpdateTag updates the label for a tag.
+func (l *Lidarr) UpdateTag(tagID int, label string) (int, error) {
+	body, err := json.Marshal(&starr.Tag{Label: label, ID: tagID})
+	if err != nil {
+		return 0, fmt.Errorf("json.Marshal(tag): %w", err)
+	}
+
+	var tag starr.Tag
+	if err = l.PutInto("v1/tag", nil, body, &tag); err != nil {
+		return tag.ID, fmt.Errorf("api.Put(tag): %w", err)
+	}
+
+	return tag.ID, nil
+}
